Initialize UTXO coin with a pubkey requested from a Trezor device * Move `UtxoCoinBuilder`, `UtxoArcBuilder` from `utxo.rs` to a separate module * Add `extended_pubkey`, `derivation_path` to `HDWalletInfo`
