+++ mm2src/coins_activation/src/platform_coin_with_tokens.rs
@@ -184,7 +184,7 @@ pub enum EnablePlatformCoinWithTokensError {
         ticker: String,
         protocol: CoinProtocol,
     },
-    #[display(fmt = "Error {} on platform coin {} creation", error, ticker)]
+    #[display(fmt = "Error on platform coin {} creation: {}", ticker, error)]
     PlatformCoinCreationError {
         ticker: String,
         error: String,
