+++ mm2src/coins_activation/src/utxo_activation/init_utxo_standard_activation_error.rs
@@ -1,5 +1,5 @@
 use crate::standalone_coin::InitStandaloneCoinError;
-use coins::utxo::UtxoCoinBuildError;
+use coins::utxo::utxo_builder::UtxoCoinBuildError;
 use coins::RegisterCoinError;
 use crypto::CryptoInitError;
 use derive_more::Display;
