+++ mm2src/coins_activation/src/utxo_activation/mod.rs
@@ -2,5 +2,6 @@ mod init_utxo_standard_activation;
 mod init_utxo_standard_activation_error;
 mod init_utxo_standard_statuses;
 mod utxo_standard_activation_result;
+mod utxo_standard_coin_hw_ops;
 
 pub use init_utxo_standard_activation::UtxoStandardTaskManagerShared;
