+++ mm2src/hw_common/src/primitives.rs
@@ -1,6 +1,8 @@
 pub const HARDENED_PATH: u32 = 2147483648;
 
-pub use bip32::{ChildNumber, DerivationPath, Error as Bip32Error};
+pub use bip32::{ChildNumber, DerivationPath, Error as Bip32Error, ExtendedPublicKey};
+
+pub type Secp256k1ExtendedPublicKey = ExtendedPublicKey<secp256k1::PublicKey>;
 
 #[derive(Clone, Copy)]
 pub enum EcdsaCurve {
