--- mm2src/trezor/src/constants.rs
@@ -1,8 +1,0 @@
-use derive_more::Display;
-
-#[derive(Clone, Copy, Debug, Display, Deserialize, Serialize)]
-pub enum TrezorCoin {
-    Bitcoin,
-    Komodo,
-    Qtum,
-}
