+++ mm2src/trezor/src/utxo/utxo_command.rs
@@ -1,7 +1,7 @@
 use crate::client::TrezorSession;
-use crate::constants::TrezorCoin;
 use crate::proto::messages_bitcoin as proto_bitcoin;
 use crate::result_handler::ResultHandler;
+use crate::utxo::TrezorUtxoCoin;
 use crate::{ecdsa_curve_to_string, serialize_derivation_path, TrezorResponse, TrezorResult};
 use hw_common::primitives::{DerivationPath, EcdsaCurve};
 
@@ -10,7 +10,7 @@ impl<'a> TrezorSession<'a> {
     pub async fn get_utxo_address<'b>(
         &'b mut self,
         path: DerivationPath,
-        coin: TrezorCoin,
+        coin: TrezorUtxoCoin,
         show_display: bool,
     ) -> TrezorResult<TrezorResponse<'a, 'b, String>> {
         let req = proto_bitcoin::GetAddress {
@@ -28,7 +28,7 @@ impl<'a> TrezorSession<'a> {
     pub async fn get_public_key<'b>(
         &'b mut self,
         path: DerivationPath,
-        coin: TrezorCoin,
+        coin: TrezorUtxoCoin,
         ecdsa_curve: EcdsaCurve,
     ) -> TrezorResult<TrezorResponse<'a, 'b, String>> {
         let req = proto_bitcoin::GetPublicKey {
